@cutecr 2021-08-17
Tested, works
```
./install.sh --nginx_option 1 --php_option 9 --phpcache_option 1 --php_extensions ioncube,imagick,fileinfo --db_option 5 --dbinstallmethod 1 --dbrootpwd $(openssl rand -hex 8) --iptables
```

```
# nginx -v
nginx version: nginx/1.20.1

# php -v
PHP 7.4.21 (cli) (built: Aug 17 2021 11:06:51) ( NTS )
Copyright (c) The PHP Group
Zend Engine v3.4.0, Copyright (c) Zend Technologies
    with the ionCube PHP Loader + ionCube24 v10.4.5, Copyright (c) 2002-2020, by ionCube Ltd.
    with Zend OPcache v7.4.21, Copyright (c), by Zend Technologies

# mysql -V
mysql  Ver 15.1 Distrib 10.6.3-MariaDB, for Linux (x86_64) using readline 5.1
```

```
# systemctl status nginx
● nginx.service - nginx - high performance web server
     Loaded: loaded (/lib/systemd/system/nginx.service; enabled; vendor preset: enabled)
     Active: active (running) since Tue 2021-08-17 10:52:27 CST; 24min ago
       Docs: http://nginx.org/en/docs/
   Main PID: 48628 (nginx)
      Tasks: 2 (limit: 2341)
     Memory: 32.4M
        CPU: 50ms
     CGroup: /system.slice/nginx.service
             ├─48628 nginx: master process /usr/local/nginx/sbin/nginx -c /usr/local/nginx/conf/nginx.conf
             └─48630 nginx: worker process

Aug 17 10:52:26 debian systemd[1]: Starting nginx - high performance web server...
Aug 17 10:52:26 debian nginx[48626]: nginx: the configuration file /usr/local/nginx/conf/nginx.conf syntax is ok
Aug 17 10:52:26 debian nginx[48626]: nginx: configuration file /usr/local/nginx/conf/nginx.conf test is successful
Aug 17 10:52:27 debian systemd[1]: Started nginx - high performance web server.

# systemctl status php-fpm
● php-fpm.service - The PHP FastCGI Process Manager
     Loaded: loaded (/lib/systemd/system/php-fpm.service; enabled; vendor preset: enabled)
     Active: active (running) since Tue 2021-08-17 11:08:01 CST; 9min ago
       Docs: http://php.net/docs.php
    Process: 297139 ExecReload=/bin/kill -USR2 $MAINPID (code=exited, status=0/SUCCESS)
   Main PID: 267199 (php-fpm)
      Tasks: 23 (limit: 2341)
     Memory: 24.1M
        CPU: 168ms
     CGroup: /system.slice/php-fpm.service
             ├─267199 php-fpm: master process (/usr/local/php/etc/php-fpm.conf)
             ├─297150 php-fpm: pool www
             ├─297151 php-fpm: pool www
             ├─297152 php-fpm: pool www
             ├─297153 php-fpm: pool www
             ├─297157 php-fpm: pool www
             ├─297158 php-fpm: pool www
             ├─297159 php-fpm: pool www
             ├─297160 php-fpm: pool www
             ├─297161 php-fpm: pool www
             ├─297162 php-fpm: pool www
             ├─297163 php-fpm: pool www
             ├─297164 php-fpm: pool www
             ├─297165 php-fpm: pool www
             ├─297166 php-fpm: pool www
             ├─297167 php-fpm: pool www
             ├─297168 php-fpm: pool www
             ├─297169 php-fpm: pool www
             ├─297170 php-fpm: pool www
             ├─297171 php-fpm: pool www
             ├─297172 php-fpm: pool www
             ├─297173 php-fpm: pool www
             └─297174 php-fpm: pool www

Aug 17 11:08:01 debian systemd[1]: Started The PHP FastCGI Process Manager.
Aug 17 11:15:18 debian systemd[1]: Reloading The PHP FastCGI Process Manager.
Aug 17 11:15:18 debian systemd[1]: Reloaded The PHP FastCGI Process Manager.

# systemctl status mysqld
● mysqld.service - LSB: start and stop MariaDB
     Loaded: loaded (/etc/init.d/mysqld; generated)
     Active: active (running) since Tue 2021-08-17 11:15:18 CST; 2min 23s ago
       Docs: man:systemd-sysv-generator(8)
    Process: 296665 ExecStart=/etc/init.d/mysqld start (code=exited, status=0/SUCCESS)
      Tasks: 11 (limit: 2341)
     Memory: 84.2M
        CPU: 455ms
     CGroup: /system.slice/mysqld.service
             ├─296854 /bin/sh /usr/local/mariadb/bin/mysqld_safe --datadir=/data/mariadb --pid-file=/data/mariadb/mysql.pid
             └─297122 /usr/local/mariadb/bin/mariadbd --basedir=/usr/local/mariadb --datadir=/data/mariadb --plugin-dir=/usr/local/mariadb/lib/plugin --user=mysql --log-error=/data/mariadb/mysql-error.log>

Aug 17 11:15:17 debian systemd[1]: Starting LSB: start and stop MariaDB...
Aug 17 11:15:17 debian mysqld[296665]: Starting MariaDB
Aug 17 11:15:17 debian mysqld[296665]: .
Aug 17 11:15:17 debian mysqld[296854]: 210817 11:15:17 mysqld_safe Logging to '/data/mariadb/mysql-error.log'.
Aug 17 11:15:17 debian mysqld[296854]: 210817 11:15:17 mysqld_safe Starting mariadbd daemon with databases from /data/mariadb
Aug 17 11:15:18 debian mysqld[296665]: .
Aug 17 11:15:18 debian systemd[1]: Started LSB: start and stop MariaDB.
```